refactor (chart): extract event hook + event sidebar
